requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
av==15.1.0
bitsandbytes==0.41.0
black==26.1.0
certifi==2025.11.12
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
DataProperty==1.1.0
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.3.1
fastapi==0.125.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
gradio_client==0.2.9
h11==0.14.0
hf-xet==1.2.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.36.0
identify==2.6.16
idna==3.11
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonlines==4.0.0
latex2mathml==3.78.1
latex2sympy2==1.9.1
loguru==0.7.3
lxml==6.0.2
markdown2==2.5.4
MarkupSafe==2.1.5
mbstrdecoder==1.1.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
nodeenv==1.10.0
numexpr==2.14.1
numpy==1.26.1
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile==1.15.0.42
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.2.0
openai==2.14.0
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.90
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
pathvalidate==3.3.1
peft==0.4.0
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.0
py-cpuinfo==9.0.0
pyarrow==23.0.0
pyarrow-hotfix==0.7
pybind11==3.0.1
pycocoevalcap==1.2
pycocotools==2.0.11
pydantic==1.10.8
pydantic_core==2.41.5
Pygments==2.19.2
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2025.11.3
requests==2.32.5
sacrebleu==2.6.0
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.15.3
sentence-transformers==5.2.0
sentencepiece==0.1.99
sentry-sdk==2.50.0
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.50.0
svgwrite==1.4.3
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.15.2
tomli==2.4.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers @ git+https://github.com/huggingface/transformers.git@1c39974a4c4036fd641bc1191cc32799f85715a4
transformers-stream-generator==0.0.5
triton==3.5.1
typeguard==4.4.4
typepy==1.3.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.5
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.40.0
virtualenv==20.36.1
wandb==0.24.0
wavedrom==2.0.3.post3
wcwidth==0.3.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.12.8
zss==1.2.0
zstandard==0.25.0
